+++ src/Xteq5Engine/AssetScriptRunner.cs
@@ -15,11 +15,11 @@ namespace Xteq5
     {
         List<AssetRecord> _results;
 
-        internal async Task<List<AssetRecord>> Run(PSScriptRunner ScriptRunner, string AssetScriptPath)
+        internal async Task<List<AssetRecord>> Run(PSScriptRunner ScriptRunner, string AssetScriptPath, IProgress<RunnerProgress> Progress = null)
         {
             _results = new List<AssetRecord>();
 
-            await RunInternalAsync(ScriptRunner, AssetScriptPath);
+            await RunInternalAsync(ScriptRunner, AssetScriptPath, Progress);
 
             return _results;
         }